requirements.txt
flask
flask_migrate
flask_login
flask_admin
flasgger
flask_sqlalchemy
flask_wtf
stripe
validators
pymysql
email_validator
cryptography
gunicorn